+++ lib/payu_pl/endpoints.rb
@@ -8,6 +8,10 @@ module PayuPl
 
     ORDERS = "/api/v2_1/orders"
 
+    SHOPS = "/api/v2_1/shops"
+    PAYOUTS = "/api/v2_1/payouts"
+    REPORTS = "/api/v2_1/reports"
+
     def self.order(order_id)
       "#{ORDERS}/#{URI.encode_www_form_component(order_id.to_s)}"
     end
@@ -26,6 +30,18 @@ module PayuPl
 
     def self.order_refund(order_id, refund_id)
       "#{order_refunds(order_id)}/#{URI.encode_www_form_component(refund_id.to_s)}"
+    end
+
+    def self.shop(shop_id)
+      "#{SHOPS}/#{URI.encode_www_form_component(shop_id.to_s)}"
+    end
+
+    def self.payout(payout_id)
+      "#{PAYOUTS}/#{URI.encode_www_form_component(payout_id.to_s)}"
+    end
+
+    def self.report(report_id)
+      "#{REPORTS}/#{URI.encode_www_form_component(report_id.to_s)}"
     end
   end
 end